attach disks when running integration tests
The vagrant instances run for integration tests are attached a 10GB thin
provisioned disk. It will show on /dev/sdb an allow for tests when
creating an OSD on a physical device.

This feature has been added to a fork of rspec-system (
https://gitorious.org/rspec-system/rspec-system/commit/6b1860fc5ab98337aa6e8aafac2ca2679f11ee47
) which is hosted on gitorious instead of github. There is no hope of
getting the patch accepted because rspec-system is retired.

The test instances are renamed (first.vm => first, second.vm => second)
without a dot in their name because it makes it easier to derive a disk
name that does not conflict with the restrictions imposed by vagrant on
disk file names.

Change-Id: I1964c7f6160bd16b92a7e4fd6dc9444c22cfd60b
Signed-off-by: Loic Dachary <loic@dachary.org>